KVM: arm64: nVHE: Mark shared variables as externally visible
There are several variables exported from nVHE which are used in the kernel
prior to when the hypervisor starts. Due to the separate linking step for
nVHE, GCC LTO causes these variables to be incorrectly optimized out.

Mark these variables as externally visible so they're not optimized out by
GCC LTO.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>